vlm/requirements.txt
fastapi
gdown==4.5.1
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opencv-contrib-python
git+https://github.com/openai/CLIP.git 
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.3.0+cu121
torchvision==0.18.0+cu121
transformers==4.37.0
uvicorn[standard]
ultralytics==8.2.27
Pillow
dill
numpy
